requirements.txt
torch>=2.7.0
torchaudio>=2.7.0
numpy>=1.23
librosa>=0.10.0
sentencepiece>=0.1.99
PyYAML>=6.0

